# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DSA Studio API
  description: REST API for the DSA Studio learning platform — through Phase 4 Advanced Features.
  version: 0.4.0
  contact:
    name: DSA Studio Team
servers:
  - url: http://localhost:4000/api
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /topics:
    get:
      tags: [Topics]
      summary: List topics
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [basic, intermediate, advanced]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated topic list
...
  /topics/{id}:
    get:
      tags: [Topics]
      summary: Get topic by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Topic detail
        '404':
          description: Topic not found
  /topics/{id}/questions:
    get:
      tags: [Topics]
      summary: List questions for a topic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Questions in topic
  /questions:
    get:
      tags: [Questions]
      summary: List questions with filters
      parameters:
        - name: topicId
          in: query
          schema:
            type: string
            format: uuid
        - name: topicSlug
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [basic, intermediate, advanced]
        - name: tag
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated question list
  /questions/{id}:
    get:
      tags: [Questions]
      summary: Get question detail (sample test cases only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Question detail
        '404':
          description: Question not found
...